refactor: move import words to controller
