+++ src/store/Api.ts
@@ -1,43 +1,8 @@
-import ajv from "ajv"
-import ajvFormats from "ajv-formats"
 import { ipcRenderer } from "electron"
-import fs from "fs"
 import { WordsController } from "./Controllers/Words"
+import { StoreOptions } from "./Schemas"
 import { options } from "./Store"
-import { StoreOptions, StoreWord } from "./Schemas"
 
-async function importWords(palavras: StoreWord[]) {
-    const validator = ajvFormats(new ajv()).compile({
-        type: "array",
-        default: [],
-        items: {
-            type: "object",
-            required: ["palavra", "definicao", "registro"],
-            properties: {
-                palavra: {
-                    type: "string"
-                },
-                definicao: {
-                    type: "string"
-                },
-                registro: {
-                    type: "string",
-                    format: "date-time"
-                },
-                ultimaEdicao: {
-                    type: "string",
-                    format: "date-time"
-                }
-            }
-        }
-    })
-
-    if (validator(palavras)) {
-        WordsController.MergeWords(palavras)
-    } else {
-        throw new Error("Arquivo inválido")
-    }
-}
 
 export const api = {
     get version() {
@@ -69,25 +34,7 @@ export const api = {
     },
 
     importWords() {
-        try {
-            const file = ipcRenderer.sendSync("get-file")
-            console.log(file)
-
-            if (file === "canceled") {
-                return "canceled"
-            }
-
-            const jsonString = fs.readFileSync(file).toString()
-            console.log(jsonString)
-
-            const json = JSON.parse(jsonString)
-            console.log(json)
-
-            importWords(json)
-        } catch (error) {
-            console.error(error)
-            return false
-        }
+        WordsController.ImportWords()
     },
 
     toggleDarkMode() {
